changed the signature of the preprocessing rules, yet to complete preprocessing rules
