+++ networkx/algorithms/vertex_covering/vertex_cover.py
@@ -2,6 +2,7 @@
 Function to find vertex cover of size atmost k
 """
 
+from networkx.algorithms.vertex_covering.lp_decomposition import find_lp_decomposition
 from networkx.algorithms.vertex_covering.vertex_cover_preprocessing import *
 from networkx.utils.decorators import not_implemented_for
 
@@ -25,6 +26,7 @@ __all__ = [
     "vertex_cover_branching",
     "max_degree_branching",
     "vc_above_lp_branching",
+    "vertex_cover",
 ]
 
 
@@ -35,29 +37,72 @@ def vertex_cover_preprocessing(G, k, vc, rules=None):
     if rules is None:
         rules = preprocessing_rules
 
-    while k > 0:
-        applied = False
+    applied = False
 
-        for rule in rules:
-            if not applied:
-                applied, G, k, to_remove_from_vc, to_add_to_vc, is_k_vc_possible = rule(
-                    G, k, vc
-                )
-
-                if not is_k_vc_possible:
-                    is_k_vc_possible = False
-                    return G, k, vc, is_k_vc_possible
-
-                if to_remove_from_vc:
-                    assert isinstance(vc, set)
-                    vc.difference_update(to_remove_from_vc)
-                if to_add_to_vc:
-                    vc.update(to_add_to_vc)
-
+    for rule in rules:
         if not applied:
-            break
+            print(rule.__name__)
+            (
+                applied,
+                G,
+                k,
+                is_k_vc_possible,
+                function_to_be_applied,
+            ) = rule(G, k, vc)
+
+            if not is_k_vc_possible:
+                is_k_vc_possible = False
+                return G, k, vc, is_k_vc_possible
+
+            # if to_remove_from_vc:
+            #     assert isinstance(vc, set)
+            #     vc.difference_update(to_remove_from_vc)
+            # if to_add_to_vc:
+            #     vc.update(to_add_to_vc)
+
+            if applied and function_to_be_applied is not None:
+                is_k_vc_exists, vc = vertex_cover(G, k)
+                function_to_be_applied(is_k_vc_exists, vc)
+
+    if applied:
+        return vertex_cover_preprocessing(G, k, vc)
 
     return G, k, vc, is_k_vc_possible
+    #
+    # while k > 0:
+    #     applied = False
+    #
+    #     for rule in rules:
+    #         if not applied:
+    #             (
+    #                 applied,
+    #                 G,
+    #                 k,
+    #                 to_remove_from_vc,
+    #                 to_add_to_vc,
+    #                 is_k_vc_possible,
+    #                 function_to_be_applied,
+    #             ) = rule(G, k, vc)
+    #
+    #             if not is_k_vc_possible:
+    #                 is_k_vc_possible = False
+    #                 return G, k, vc, is_k_vc_possible
+    #
+    #             # if to_remove_from_vc:
+    #             #     assert isinstance(vc, set)
+    #             #     vc.difference_update(to_remove_from_vc)
+    #             # if to_add_to_vc:
+    #             #     vc.update(to_add_to_vc)
+    #
+    #             if applied and function_to_be_applied is not None:
+    #                 is_k_vc_exists, vc = vertex_cover(G, k)
+    #                 function_to_be_applied(is_k_vc_exists, vc)
+    #
+    #     if not applied:
+    #         break
+    #
+    # return G, k, vc, is_k_vc_possible
+    #
 
 
 @not_implemented_for("directed")
@@ -140,3 +185,37 @@ def max_degree_branching(G, k):
 def vc_above_lp_branching(G, k):
     rules = vc_above_lp_branch_preprocessing_rules
     return vertex_cover_branching(G, k, rules)
+
+
+def vertex_cover(G, k):
+    # find lp-opt value
+    # compare 1.4656^k and 2.618^(k - lpOpt)
+    if len(G) == 0:
+        return True, set()
+
+    g_new = G.copy()
+
+    vc = set()
+    g_new, k, vc, is_k_vc_possible = vertex_cover_preprocessing(g_new, k, vc)
+
+    if not is_k_vc_possible:
+        return False, set()
+
+    lp_opt_value, *_ = find_lp_decomposition(g_new)
+
+    if lp_opt_value > k:
+        return False, set()
+
+    vc_above_lp_opt_algo_check = 2.618 ** (k - lp_opt_value)
+    max_deg_algo_check = 1.4656**k
+
+    is_k_vc_exists, vc_sub = False, set()
+
+    if vc_above_lp_opt_algo_check < max_deg_algo_check:
+        is_k_vc_exists, vc_sub = vc_above_lp_branching(g_new, k)
+    else:
+        is_k_vc_exists, vc_sub = max_degree_branching(g_new, k)
+
+    vc.update(vc_sub)
+
+    return is_k_vc_exists, vc
